update user comp, unfinished
